REFACTOR diatonic functions to their own file, moved ExpChord to Base.Chord.Chord for now, and moved notesToChord to Lib
